Fix integer overflow in delta source size
delta-source-raw-count is read into an int and multiplied by
FWUP_BLOCK_SIZE, so the product overflows for any source region of 2 GiB
or more. At 4 GiB - 8388608 blocks, a common A/B rootfs partition size -
the top bit is lost and xdelta3 is handed a zero-length source, failing
every delta with "xdelta3 error: source file too short".

Both the count and xd_source_count become int64_t. Widening the local
alone is not enough: xd_source_count was size_t, which is 32 bits on
32-bit targets and would truncate the result again on assignment.
off_t is not usable either - it is 32 bits on mingw-w64 unless
_FILE_OFFSET_BITS=64 precedes <sys/types.h>, and functions.h includes
that header without config.h.

The source-region bounds check compared off_t against size_t; it is now
signed and fixed-width throughout.

+++ src/functions.h
@@ -19,6 +19,7 @@
 
 #include <stddef.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <sys/types.h>
 #include <confuse.h>
 #include "uboot_env.h"
@@ -66,7 +67,7 @@ struct fun_context {
     // Delta update processing (NULL if not in use)
     struct xdelta_state *xd;
     off_t xd_source_offset;
-    size_t xd_source_count;
+    int64_t xd_source_count;
     const char *xd_source_path;
     struct disk_crypto *xd_source_dc;
